[ALPS04859329] apusys power : change power_bit_mask to 32bits
[Detail]
1. change power_bit_mask to 32 bits

MTK-Commit-Id: 777b0862c5cdffaabbe19092ade50b20c2900b16

CR-Id: ALPS04859329
Feature: [Module] APUSYS

+++ drivers/misc/mediatek/apusys/power/2.0/mt6885/apusys_power_cust.h
@@ -132,7 +132,7 @@ struct apusys_dvfs_opps {
 	uint8_t user_opp_index[APUSYS_DVFS_USER_NUM];
 	uint8_t driver_opp_index[APUSYS_DVFS_USER_NUM];
 	bool is_power_on[APUSYS_POWER_USER_NUM];
-	uint8_t power_bit_mask;
+	uint32_t power_bit_mask;
 	uint64_t id;
 	enum DVFS_VOLTAGE vsram_volatge;
 };
